+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add support for ccminer-xevan
 * Don't require a restart when changing the region
 * Re-download miner if deleted during mining
+* Improve First Run Wizard
 * Update ethminer to v0.14.0.dev4
 
 ## RudeHash 5.0 (2018-03-09)